+++ src/propertyFields/propertyEditor/IPropertyPanePropertyEditorHost.ts
@@ -1,0 +1,15 @@
+/**
+ * PropertyFieldColorPickerHost properties interface
+ */
+export interface IPropertyPanePropertyEditorHostProps {
+	webpart: any;
+}
+
+/**
+ * PropertyFieldColorPickerHost state interface
+ */
+export interface IPropertyPanePropertyEditorHostState {
+	errorMessage?: string;
+	openPanel? : boolean;
+	propertiesJson?: string;
+}